Move tests to component and pages folders
