+++ pages/__tests__/index.text.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { render, screen } from "@testing-library/react"
 import "@testing-library/jest-dom"
-import Home from "../../pages/index"
+import Home from "../index"
 
 describe("Home", () => {
   it("should render an H2", () => {
